replace token.decode with getTokenDecoder
This feature is the ability to get information out of the token if one
is provided and not fail if one if not provided. Before this change, it
was not possible to get information out of a token and not fail if one
was not provided. Now you can decide per route if a token is _required_ to
consume the route. Flexibility good.